+++ changelog.md
@@ -14,12 +14,19 @@
 - [ ] Add anti dupe system with NMS Tag 
 - [ ] Add boolean for enable or disable inventory to load
 - [ ] Add nbt support
+- [ ] Upgrade potion support for multiple effect, main effect and color
 
 # Unreleased
 
+- Added number of inventory loaded with command /zm reload
 - Fix model id value in config, you can now you modelId, customModelId or customModelData
 - Fix player head for 1.20+ (https://blog.jeff-media.com/creating-custom-heads-in-spigot-1-18-1/) https://github.com/Maxlego08/zMenu/issues/16
-- Added number of inventory loaded with order/zm reload
+- Fix docs for potion
+- Added information about potion in 1.8 - 1.12. You have to do this:
+```yml
+material: POTION
+durability: 16454
+```
 
 # 1.0.1.6